Add new hooks for row node handling and enhance display options
- Introduced `getNodeLabel`, `getComponentTypeIcon`, and `getNodeTypeColor` functions to manage node labeling and icon representation based on workspace context and node types.
- Created `useRowNodeActions` and `useRowNodeDisplay` hooks to encapsulate logic for managing row actions and display properties, improving code organization and reusability.
- Added `row-node-label.ts` and `use-row-node-actions.ts` files to implement the new functionality, enhancing the sidebar's interactivity and user experience.
- Updated `use-row-node.ts` to integrate the new hooks, streamlining the row rendering process and improving maintainability.

These changes significantly enhance the sidebar's functionality by providing more intuitive node interactions and display options, improving overall user experience.
